refactor: 범용적인 유틸 파일을 utils 폴더 하위로 이동
- Validate.js 를 이동
- RacingcarManager.js 에서 getRandomNumer()를 Random.js 로 분리
