+++ src/utils/Random.js
@@ -1,0 +1,5 @@
+function getRandomNumber() {
+    return Math.floor(Math.random() * 10);
+}
+
+export default getRandomNumber;
